refactoring the tools module to add user id

+++ src/modules/tools/services/DeleteToolService.ts
@@ -4,6 +4,7 @@ import IToolsRepository from '../repositories/IToolsRepository';
 
 interface IRequest {
   id: string;
+  user_id: string;
 }
 
 @injectable()
@@ -13,11 +14,15 @@ export default class DeleteToolService {
     private toolsRepository: IToolsRepository,
   ) {}
 
-  public async execute({ id }: IRequest): Promise<void> {
+  public async execute({ id, user_id }: IRequest): Promise<void> {
     const tool = await this.toolsRepository.findOneById(id);
 
     if (!tool) {
       throw new AppError('Tool not found', 404);
+    }
+
+    if (tool.user_id !== user_id) {
+      throw new AppError("You don't can delete this tool, only owner.");
     }
 
     await this.toolsRepository.delete(tool);
